+++ .github/workflows/build.yml
@@ -22,7 +22,7 @@ jobs:
 
       - name: Install Dependencies
         run: |
-          python -m pip install -U jupyterlab~=4.0 jupyter_packaging~=0.10
+          python -m pip install -U jupyterlab~=4.0 jupyter_packaging~=0.10 "notebook<7"
 
       - name: Install the Voilà Preview JupyterLab extension
         run: |
